requirements.txt
langchain-openai
tiktoken
chromadb
langchain
gitpython
python-dotenv
